requirements.txt
Django==1.9.2
django-braces>=1.8

